Remove generator script's list of ignored syscalls
This list was originally intended to prevent exporting `SYS_*` symbols
for syscalls that are not part of the Linux kernel's current ABI,
because they're internal (`restart_syscall`), deprecated (`remap_file_pages`),
or removed (`nfsservctl`). But in practice this isn't important (users
could always just `syscall!()` integers directly), and it introduces
unnecessary deviation between the kernel sources and this library.

GitOrigin-RevId: c88fbd848c053e6bc301c88e41e5074d01554630